Get all services separately
