+++ x-pack/legacy/plugins/apm/server/lib/service_map/get_service_map.ts
@@ -3,29 +3,18 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import { uniq, take, sortBy } from 'lodash';
+
 import { PromiseReturnType } from '../../../typings/common';
-import { rangeFilter } from '../helpers/range_filter';
 import {
   Setup,
   SetupTimeRange,
   SetupUIFilters
 } from '../helpers/setup_request';
-import {
-  SERVICE_NAME,
-  SERVICE_ENVIRONMENT,
-  DESTINATION_ADDRESS,
-  PROCESSOR_EVENT,
-  PARENT_ID,
-  TRANSACTION_TYPE,
-  TRANSACTION_NAME,
-  SPAN_TYPE,
-  SPAN_SUBTYPE,
-  TRACE_ID,
-  TRANSACTION_SAMPLED
-} from '../../../common/elasticsearch_fieldnames';
-import { ESFilter } from '../../../typings/elasticsearch';
 import { getServiceMapFromTraceIds } from './get_service_map_from_trace_ids';
+import { getTraceSampleIds } from './get_trace_sample_ids';
+import { getServicesProjection } from '../../../common/projections/services';
+import { mergeProjection } from '../../../common/projections/util/merge_projection';
+import { SERVICE_AGENT_NAME } from '../../../common/elasticsearch_fieldnames';
 
 export interface IEnvOptions {
   setup: Setup & SetupTimeRange & SetupUIFilters;
@@ -34,168 +23,18 @@ export interface IEnvOptions {
   after?: string;
 }
 
-const MAX_TRACES_TO_INSPECT = 1000;
-
-export type ServiceMapAPIResponse = PromiseReturnType<typeof getServiceMap>;
-export async function getServiceMap({
+async function getConnectionData({
   setup,
   serviceName,
   environment,
   after
 }: IEnvOptions) {
-  const isTop = !after;
-  const isAll = !serviceName;
-
-  const { indices, start, end, client, uiFiltersES } = setup;
-
-  let sampleIndices = [indices['apm_oss.spanIndices']];
-
-  let processorEvent = 'span';
-
-  const rangeQuery = { range: rangeFilter(start, end) };
-
-  if (isTop) {
-    sampleIndices = [indices['apm_oss.transactionIndices']];
-    processorEvent = 'transaction';
-  }
-
-  const query = {
-    bool: {
-      filter: [
-        { term: { [PROCESSOR_EVENT]: processorEvent } },
-        rangeQuery,
-        ...uiFiltersES
-      ] as ESFilter[]
-    }
-  } as { bool: { filter: ESFilter[]; must_not?: ESFilter[] | ESFilter } };
-
-  if (serviceName) {
-    query.bool.filter.push({ term: { [SERVICE_NAME]: serviceName } });
-  }
-
-  if (environment) {
-    query.bool.filter.push({ term: { [SERVICE_ENVIRONMENT]: environment } });
-  }
-
-  if (isTop && isAll) {
-    query.bool.must_not = { exists: { field: PARENT_ID } };
-  }
-
-  if (processorEvent === 'transaction') {
-    query.bool.filter.push({ term: { [TRANSACTION_SAMPLED]: true } });
-  }
-
-  const afterObj =
-    after && after !== 'top'
-      ? { after: JSON.parse(Buffer.from(after, 'base64').toString()) }
-      : {};
-
-  const params = {
-    index: sampleIndices,
-    body: {
-      size: 0,
-      query,
-      aggs: {
-        connections: {
-          composite: {
-            size: 1000,
-            ...afterObj,
-            sources: [
-              { [SERVICE_NAME]: { terms: { field: SERVICE_NAME } } },
-              {
-                [SERVICE_ENVIRONMENT]: {
-                  terms: { field: SERVICE_ENVIRONMENT, missing_bucket: true }
-                }
-              },
-              {
-                [TRANSACTION_TYPE]: {
-                  terms: { field: TRANSACTION_TYPE, missing_bucket: true }
-                }
-              },
-              {
-                [TRANSACTION_NAME]: {
-                  terms: { field: TRANSACTION_NAME, missing_bucket: true }
-                }
-              },
-              {
-                [SPAN_TYPE]: {
-                  terms: { field: SPAN_TYPE, missing_bucket: true }
-                }
-              },
-              {
-                [SPAN_SUBTYPE]: {
-                  terms: { field: SPAN_SUBTYPE, missing_bucket: true }
-                }
-              },
-              {
-                [DESTINATION_ADDRESS]: {
-                  terms: { field: DESTINATION_ADDRESS, missing_bucket: true }
-                }
-              }
-            ]
-          },
-          aggs: {
-            trace_id_samples: {
-              diversified_sampler: {
-                shard_size: 20,
-                ...(processorEvent === 'transaction'
-                  ? { field: TRACE_ID }
-                  : {
-                      script: {
-                        lang: 'painless',
-                        source: `(int)doc['span.duration.us'].value/100000`
-                      }
-                    })
-              },
-              aggs: {
-                sample_documents: {
-                  top_hits: {
-                    size: 20,
-                    _source: ['trace.id']
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  };
-
-  const tracesSampleResponse = await client.search<
-    { trace: { id: string } },
-    typeof params
-  >(params);
-
-  let nextAfter: string | undefined;
-
-  const receivedAfterKey =
-    tracesSampleResponse.aggregations?.connections.after_key;
-
-  if (!after) {
-    nextAfter = 'top';
-  } else if (receivedAfterKey) {
-    nextAfter = Buffer.from(JSON.stringify(receivedAfterKey)).toString(
-      'base64'
-    );
-  }
-
-  // make sure at least one trace per composite/connection bucket
-  // is queried
-  const traceIdsWithPriority =
-    tracesSampleResponse.aggregations?.connections.buckets.flatMap(bucket =>
-      bucket.trace_id_samples.sample_documents.hits.hits.map((hit, index) => ({
-        traceId: hit._source.trace.id,
-        priority: index
-      }))
-    ) || [];
-
-  const traceIds = take(
-    uniq(
-      sortBy(traceIdsWithPriority, 'priority').map(({ traceId }) => traceId)
-    ),
-    MAX_TRACES_TO_INSPECT
-  );
+  const { traceIds, after: nextAfter } = await getTraceSampleIds({
+    setup,
+    serviceName,
+    environment,
+    after
+  });
 
   const serviceMapData = traceIds.length
     ? await getServiceMapFromTraceIds({
@@ -209,5 +48,65 @@ export async function getServiceMap({
   return {
     after: nextAfter,
     ...serviceMapData
+  };
+}
+
+async function getServicesData(options: IEnvOptions) {
+  // only return services on the first request for the global service map
+  if (options.serviceName || options.after) {
+    return [];
+  }
+
+  const { setup } = options;
+
+  const projection = getServicesProjection({ setup });
+
+  const params = mergeProjection(projection, {
+    body: {
+      size: 0,
+      aggs: {
+        services: {
+          terms: {
+            field: projection.body.aggs.services.terms.field,
+            size: 500
+          },
+          aggs: {
+            agent_name: {
+              terms: {
+                field: SERVICE_AGENT_NAME
+              }
+            }
+          }
+        }
+      }
+    }
+  });
+
+  const { client } = setup;
+
+  const response = await client.search(params);
+
+  return (
+    response.aggregations?.services.buckets.map(bucket => {
+      return {
+        'service.name': bucket.key as string,
+        'agent.name':
+          (bucket.agent_name.buckets[0]?.key as string | undefined) || '',
+        'service.environment': options.environment || null
+      };
+    }) || []
+  );
+}
+
+export type ServiceMapAPIResponse = PromiseReturnType<typeof getServiceMap>;
+export async function getServiceMap(options: IEnvOptions) {
+  const [connectionData, servicesData] = await Promise.all([
+    getConnectionData(options),
+    getServicesData(options)
+  ]);
+
+  return {
+    ...connectionData,
+    services: servicesData
   };
 }
